requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
distlib==0.3.1
filelock==3.0.12
gast==0.2.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pipenv==2020.11.15
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
tf-nightly-gpu==2.5.0.dev20201111
virtualenv==20.1.0
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
yfinance==0.1.55
zipp==3.4.0

pandas==0.24
PyDrive==1.3.1
pillow==8.0.1
streamlit==0.71.0
matplotlib==3.3.3
keras==2.4.3
scipy==1.5.4
seaborn==0.11.0